backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.34.0",
    # Database
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "sqlalchemy>=2.0.0",
    # Google OAuth
    "httpx>=0.27.0",
    "pyarrow>=15.0.0",
    # Security
    "cryptography>=43.0.0",
    "greenlet>=3.3.0",
    # Database migrations
    "alembic>=1.14.0",
    # JWT authentication
    "pyjwt>=2.8.0",
    "psycopg2-binary>=2.9.11",
    "google-cloud-bigquery[bqstorage]>=3.40.0",
]

[dependency-groups]
dev = [
    "duckdb>=1.4.3",
    "pandas>=2.3.3",
]
